Date parsing that is looser and stricter.
It allows for a space between date and time in addition to T as
suggested in the text of the RFC.

It also limits months, days of the month, hours, minutes and seconds to
reasonable numbers. It doesn't yet try to parse things, but it keeps the
strings closer to real ISO Date Times.